requirements.txt
fastapi
uvicorn
langchain
langchain-google-genai
python-dotenv
langchain-core
langchain-cerebras
sarvamai